10.30.25
added inventory, loot, and stats menus

added debug developer objects and script. Dynamic hud that allows rapid testing of methods and a screen console.

reorganized level 1 scenes in hierarchy

updated the prefabs of the loots

redid the legendary item to have a parent of the vfx and sphere collider. need to do the same treatment to the other loot drops.

expanded loot behavior and add three child scripts based on it.

Modified inventory to prepare for the items

Added all the missing project tags and added ignore logic for collisions on all loot

turned off raycast target in uis

fixed mana pickup logic