Item development
items now spawn and randomize loot. Perks are randomized and the information of the two form the UI information.

Added an object hider to HUD to hide stuff until game starts.

We SUPER need to scale the hud size down and make it a dynamic hud.

Snagged a good chunk of the blink item icons and moved them into the main item folder.

Added healing efficiency perk to add to item perk loot pool.

The math on adding perks still needs to be tested. It's tested to the end of the system. must test if the values work appropriately on the player itself.

changed to scene 1. dupe scene to manually sync.

changed legendary color in color manager to match sampled color of vfx

added tooltips on utiliy tools